requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
requests>=2.31.0,<3.0.0
beautifulsoup4>=4.12.0,<5.0.0
python-dateutil>=2.8.2,<3.0.0
eval_type_backport>=0.3.1,<1.0.0
